requirements.txt
numpy>=1.23
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.7
h5py>=3.8
scipy>=1.10
tqdm>=4.65
torch>=2.0
torchvision>=0.15
tensorflow>=2.12